Moving Interval into its own module.
